Add useLocaleTime
